Token generation moved to UserServices. Made UserController method async.
