+++ backend/Lingva.BusinessLayer/Contracts/IUserService.cs
@@ -7,6 +7,7 @@ namespace Lingva.BusinessLayer.Contracts
 {
     public interface IUserService
     {
+        string GetUserToken(User user, string secret);
         User Authenticate(string username, string password);
         IEnumerable<User> GetAll();
         User GetById(int id);
