+++ backend/Lingva.DataAccessLayer/Repositories/IRepositoryUser.cs
@@ -3,7 +3,7 @@ using Lingva.DataAccessLayer.Entities;
 
 namespace Lingva.DataAccessLayer.Repositories
 {
-    public interface IUserRepository:IRepository<User>
+    public interface IRepositoryUser:IRepository<User>
     {
 
     }
